Feat: findByid includes in Produtos and Categorias

+++ ecommerce/src/main/java/com/gustavo/ecommerce/repository/PedidoRepository.java
@@ -3,5 +3,5 @@ package com.gustavo.ecommerce.repository;
 import com.gustavo.ecommerce.entity.Pedido;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PedidoRepository extends JpaRepository<Pedido, Long> {
+public interface PedidoRepository extends JpaRepository<Pedido, Integer> {
 }
